refactor(PasswordFieldset): separate NewPasswordFieldset into it's own directory
