+++ src/components/Fieldsets/PasswordFieldset/PasswordFieldset.module.scss
@@ -24,63 +24,3 @@
     border: none;
   }
 }
-
-.newPasswordInput input:global(.dark) {
-  /* Shouldn't be neccessary once everything is under the new forms, but we need to add all these specificites so the base styles don't override this. */
-  &:required[type='text'],
-  &:required[type='password'],
-  &[type='text'],
-  &[type='password'] {
-    border-width: 0;
-  }
-}
-
-
-.strengthMeter {
-  appearance: none;
-  position: absolute;
-  bottom: -0.3rem;
-  left: 0;
-
-  width: 100%;
-  height: 0.3rem;
-
-  background-image: none;
-
-  pointer-events: none;
-
-  &[value='0'] {
-    background-color: $red;
-  }
-
-  &[value='1'] {
-    background-image:
-      linear-gradient(
-        to right,
-        $red 40%,
-        $transparent 40%
-      );
-  }
-
-  &[value='2'] {
-    background-image:
-      linear-gradient(
-        to right,
-        $yellow 60%,
-        $transparent 60%
-      );
-  }
-
-  &[value='3'] {
-    background-image:
-      linear-gradient(
-        to right,
-        $yellow 80%,
-        $transparent 80%
-      );
-  }
-
-  &[value='4'] {
-    background-color: $green;
-  }
-}
