+++ src/components/Fieldsets/PasswordFieldset/index.js
@@ -1,2 +1,1 @@
 export { default } from './PasswordFieldset'
-export { default as NewPasswordFieldset } from './NewPasswordFieldset'
